+++ src/app/(sub)/auth/socialNickname/page.tsx
@@ -14,7 +14,6 @@ const SocialNicknamePage = () => {
   const [nickname, setNickname] = useState("");
   const [error, setError] = useState<string | null>(null);
   const [loading, setLoading] = useState(true);
-
   const supabase = createClient();
 
   useEffect(() => {
